Percolate command architecture; Minor bug fixes throughout (#8)
* Implement PasswordCommand

* Fix missing mask for that ONE lantern

* Fix TextInputDisplay returning the wrong text field

* Rework integer types in PlayerData; fix setting of wrong name parameter

* Percolated remaining commands into ClientConnection.cpp

* Add in SayCommand

* Correct Male/Female transposition

* Added ability to send text commands

+++ src/new_client/clientlib/PlayerData.h
@@ -70,7 +70,7 @@ public:
   void setClientShouldHideWalls( bool shouldHideWalls );
   void setOkeyName( char ( &pdataName )[ 40 ] );
 
-  long unsigned int getRaceAndSex() const;
+  int               getRaceAndSex() const;
   std::string       getPassword() const;
   OkeyPasswordType  getPasswordOkeyValues() const;
   long unsigned int getUserNumber() const;
